Fix: selection fast update on button press
Issue: When the buttons are pressed for moving the selected chess square, the selection traverses multiple squares instead of a single square.
Reason: At every positive edge of clock the button press is checked and the selected square location is updated on the display. The ClockFrequencyDivider module is instantiated to reduce the frequency and update the selection slowly so that it is controllable by the button press.