Fix: pawn moves even after invalid move
Issue: when the move is not valid the pawn doesn't move and is unlocked. But when the selection moves to the right square, the pawn moves even though it is unlocked.
Fix: check for lock before updating the position of chess piece. Otherwise when the lock switch is off the validity of the move for the last locked piece is still being checked.