Fix: black player valid move not updated on board
Issue: Even when the black makes valid move, the chess piece position is not updated.
Reason: There is a logic to check if the destination square has the piece of the same or other colour. It does so by checking the 4th bit which is 1 for white and 0 for black.
The problem is even the empty spaces without any chess piece has 0 in the 4th bit position.
Fix: Move if  the destination square has opposite colour piece of empty space.