Fix: auto reset chess layout after checkmate
Issue: When the king is captured, the whole game is automatically reset. The result is not shown on the display and the state skips to START_STATE.
Reason: The State reg variable was declared as a 2 bit number but it was actually a 3 bit number. When it transitioned from PLAY_STATE (b010) to END_STATE (b100), the State variable truncated it to first two bits (b00). This is a undefined State value and thus in the switch case the default case is triggered. The default case brings it back to START_STATE thus causing game restart.
The issue was found by examining the connection of RTL layout.
Fix: The issue was fixed by increasing the bit width of State variable from 2 to 3.
Verification: Hardware verification was performed to check if it solved the problem.
Result: Issue resolved successfully.